consolidate sia config objects into a single module
