+++ libs/go/sia/aws/agent/agent.go
@@ -38,7 +38,8 @@ import (
 	"github.com/AthenZ/athenz/libs/go/sia/access/tokens"
 	"github.com/AthenZ/athenz/libs/go/sia/aws/attestation"
 	"github.com/AthenZ/athenz/libs/go/sia/aws/options"
-	"github.com/AthenZ/athenz/libs/go/sia/aws/sds"
+	sc "github.com/AthenZ/athenz/libs/go/sia/config"
+	"github.com/AthenZ/athenz/libs/go/sia/sds"
 	"github.com/AthenZ/athenz/libs/go/sia/ssh/hostkey"
 	"github.com/AthenZ/athenz/libs/go/sia/util"
 	"github.com/ardielle/ardielle-go/rdl"
@@ -88,7 +89,7 @@ func RoleKey(rotateKey bool, roleKey, svcKey string) (*rsa.PrivateKey, error) {
 	}
 }
 
-func GetRoleCertificates(ztsUrl string, opts *options.Options) (int, []string) {
+func GetRoleCertificates(ztsUrl string, opts *sc.Options) (int, []string) {
 
 	//initialize our return state to success
 	failures := make([]string, 0)
@@ -173,7 +174,7 @@ func GetRoleCertificates(ztsUrl string, opts *options.Options) (int, []string) {
 	return len(opts.Roles), failures
 }
 
-func RegisterInstance(data []*attestation.AttestationData, ztsUrl string, opts *options.Options, docExpiryCheck bool) error {
+func RegisterInstance(data []*attestation.AttestationData, ztsUrl string, opts *sc.Options, docExpiryCheck bool) error {
 
 	//special handling for EC2 instances
 	//before we process our register event we need to check to
@@ -193,7 +194,7 @@ func RegisterInstance(data []*attestation.AttestationData, ztsUrl string, opts *
 	return nil
 }
 
-func RefreshInstance(data []*attestation.AttestationData, ztsUrl string, opts *options.Options) error {
+func RefreshInstance(data []*attestation.AttestationData, ztsUrl string, opts *sc.Options) error {
 	for i, svc := range opts.Services {
 		err := refreshSvc(svc, data[i], ztsUrl, opts)
 		if err != nil {
@@ -203,7 +204,7 @@ func RefreshInstance(data []*attestation.AttestationData, ztsUrl string, opts *o
 	return nil
 }
 
-func getServiceHostname(opts *options.Options, svc options.Service, fqdn bool) string {
+func getServiceHostname(opts *sc.Options, svc sc.Service, fqdn bool) string {
 	if !opts.SanDnsHostname {
 		return ""
 	}
@@ -237,7 +238,7 @@ func getServiceHostname(opts *options.Options, svc options.Service, fqdn bool) s
 	return fmt.Sprintf("%s.%s.%s.%s", hostname, svc.Name, hyphenDomain, opts.HostnameSuffix)
 }
 
-func registerSvc(svc options.Service, data *attestation.AttestationData, ztsUrl string, opts *options.Options) error {
+func registerSvc(svc sc.Service, data *attestation.AttestationData, ztsUrl string, opts *sc.Options) error {
 
 	key, err := util.GenerateKeyPair(2048)
 	if err != nil {
@@ -349,7 +350,7 @@ func registerSvc(svc options.Service, data *attestation.AttestationData, ztsUrl 
 	return nil
 }
 
-func refreshSvc(svc options.Service, data *attestation.AttestationData, ztsUrl string, opts *options.Options) error {
+func refreshSvc(svc sc.Service, data *attestation.AttestationData, ztsUrl string, opts *sc.Options) error {
 
 	keyFile := util.GetSvcKeyFileName(opts.KeyDir, svc.KeyFilename, opts.Domain, svc.Name)
 	certFile := util.GetSvcCertFileName(opts.CertDir, svc.CertFilename, opts.Domain, svc.Name)
@@ -462,7 +463,7 @@ func refreshSvc(svc options.Service, data *attestation.AttestationData, ztsUrl s
 	return nil
 }
 
-func generateSshRequest(opts *options.Options, primaryServiceName, hostname string) (*zts.SSHCertRequest, string, error) {
+func generateSshRequest(opts *sc.Options, primaryServiceName, hostname string) (*zts.SSHCertRequest, string, error) {
 	var err error
 	var sshCsr string
 	var sshCertRequest *zts.SSHCertRequest
@@ -556,7 +557,7 @@ func hostCertificateLinePresent(sshConfigFile, sshCertFile string) (bool, error)
 	return false, nil
 }
 
-func SetupAgent(opts *options.Options, siaAgentDir, siaLinkDir string) {
+func SetupAgent(opts *sc.Options, siaAgentDir, siaLinkDir string) {
 
 	//first, let's determine if we need to drop our privileges
 	//since it requires us to create the directories with the
@@ -613,7 +614,7 @@ func SetupAgent(opts *options.Options, siaAgentDir, siaLinkDir string) {
 	}
 }
 
-func RunAgent(siaCmds, ztsUrl string, opts *options.Options) {
+func RunAgent(siaCmds, ztsUrl string, opts *sc.Options) {
 	log.Printf("sia command line arguments specified: '%s'\n", siaCmds)
 	cmds := strings.Split(siaCmds, ",")
 	for _, cmd := range cmds {
@@ -621,7 +622,7 @@ func RunAgent(siaCmds, ztsUrl string, opts *options.Options) {
 	}
 }
 
-func runAgentCommand(siaCmd, ztsUrl string, opts *options.Options) {
+func runAgentCommand(siaCmd, ztsUrl string, opts *sc.Options) {
 
 	//make sure the meta endpoint is configured by the caller
 	if opts.MetaEndPoint == "" {
@@ -889,7 +890,7 @@ func accessTokenRequest(tokenOpts *config.TokenOptions) error {
 	return err
 }
 
-func tokenOptions(opts *options.Options, ztsUrl string) (*config.TokenOptions, error) {
+func tokenOptions(opts *sc.Options, ztsUrl string) (*config.TokenOptions, error) {
 	userAgent := fmt.Sprintf("%s-%s", opts.Provider, opts.InstanceId)
 	tokenOpts, err := tokens.NewTokenOptions(opts, ztsUrl, userAgent)
 	if err != nil {
@@ -924,7 +925,7 @@ func fetchAccessToken(tokenOpts *config.TokenOptions) error {
 	}
 }
 
-func shouldSkipRegister(opts *options.Options) bool {
+func shouldSkipRegister(opts *sc.Options) bool {
 	if opts.EC2StartTime == nil {
 		return false
 	}
@@ -933,13 +934,13 @@ func shouldSkipRegister(opts *options.Options) bool {
 	return duration.Seconds() > 1800
 }
 
-func serviceAlreadyRegistered(opts *options.Options, svc options.Service) bool {
+func serviceAlreadyRegistered(opts *sc.Options, svc sc.Service) bool {
 	keyFile := util.GetSvcKeyFileName(opts.KeyDir, svc.KeyFilename, opts.Domain, svc.Name)
 	certFile := util.GetSvcCertFileName(opts.CertDir, svc.CertFilename, opts.Domain, svc.Name)
 	return util.FileExists(keyFile) && util.FileExists(certFile)
 }
 
-func shouldExitRightAway(failedRefreshCount int, opts *options.Options) bool {
+func shouldExitRightAway(failedRefreshCount int, opts *sc.Options) bool {
 	// if the failed count already matches or exceeds our configured
 	// value then we return right away
 	if failedRefreshCount >= opts.FailCountForExit {
