+++ libs/go/sia/aws/options/options_test.go
@@ -32,9 +32,9 @@ import (
 	"testing"
 
 	"github.com/AthenZ/athenz/libs/go/sia/access/config"
+	sc "github.com/AthenZ/athenz/libs/go/sia/config"
 	"github.com/AthenZ/athenz/libs/go/sia/ssh/hostkey"
 	"github.com/AthenZ/athenz/libs/go/sia/util"
-
 	"github.com/dimfeld/httptreemux"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -88,7 +88,7 @@ func (t *testServer) httpUrl() string {
 	return fmt.Sprintf("http://%s", t.addr)
 }
 
-func getConfig(fileName, roleSuffix, metaEndPoint string, useRegionalSTS bool, region string) (*Config, *ConfigAccount, error) {
+func getConfig(fileName, roleSuffix, metaEndPoint string, useRegionalSTS bool, region string) (*sc.Config, *sc.ConfigAccount, error) {
 	// Parse config bytes first, and if that fails, load values from Instance Profile and IAM info
 	cfg, cfgAccount, err := InitFileConfig(fileName, metaEndPoint, useRegionalSTS, region, "")
 	if err != nil {
@@ -105,10 +105,10 @@ func getConfig(fileName, roleSuffix, metaEndPoint string, useRegionalSTS bool, r
 	return cfg, cfgAccount, nil
 }
 
-func getAccessProfileConfig(fileName, metaEndPoint string) (*ConfigAccount, *AccessProfileConfig, error) {
+func getAccessProfileConfig(fileName, metaEndPoint string) (*sc.ConfigAccount, *sc.AccessProfileConfig, error) {
 	// Parse config bytes first, and if that fails, load values from Instance Profile and IAM info
 	profileConfig, err := InitAccessProfileFileConfig(fileName)
-	var configAccount *ConfigAccount = nil
+	var configAccount *sc.ConfigAccount = nil
 	if err != nil {
 		log.Printf("unable to parse configuration file, error: %v\n", err)
 		log.Println("trying to determine profile name from instance profile arn...")
@@ -121,7 +121,7 @@ func getAccessProfileConfig(fileName, metaEndPoint string) (*ConfigAccount, *Acc
 	return configAccount, profileConfig, nil
 }
 
-func assertService(expected Service, actual Service) bool {
+func assertService(expected sc.Service, actual sc.Service) bool {
 	log.Printf("expected: %+v\n", expected)
 	log.Printf("actual: %+v\n", actual)
 	return expected.Name == actual.Name &&
@@ -134,7 +134,7 @@ func assertService(expected Service, actual Service) bool {
 		expected.Threshold == actual.Threshold
 }
 
-func assertInServices(svcs []Service, actual Service) bool {
+func assertInServices(svcs []sc.Service, actual sc.Service) bool {
 	log.Printf("svcs passed: %+v\n", svcs)
 	log.Printf("actual: %+v\n", actual)
 	for _, s := range svcs {
@@ -197,9 +197,9 @@ func TestOptionsNoConfig(t *testing.T) {
 	assert.Equal(t, 1, len(opts.Services))
 	assert.Equal(t, "athenz", opts.Domain)
 	assert.Equal(t, "athenz.hockey", opts.Name)
-	assert.True(t, assertService(opts.Services[0], Service{Name: "hockey", Uid: idCommandId("-u"), Gid: idCommandId("-g"), FileMode: 288, Threshold: DefaultThreshold}))
-	assert.Equal(t, DefaultThreshold, opts.Threshold)
-	assert.Equal(t, DefaultThreshold, opts.SshThreshold)
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "hockey", Uid: idCommandId("-u"), Gid: idCommandId("-g"), FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.Equal(t, sc.DefaultThreshold, opts.Threshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.SshThreshold)
 	assert.False(t, opts.FileDirectUpdate)
 	assert.Equal(t, "/var/lib/sia/keys", opts.KeyDir)
 	assert.Equal(t, "/var/lib/sia/certs", opts.CertDir)
@@ -234,8 +234,8 @@ func TestOptionsNoProfileConfig(t *testing.T) {
 	assert.True(t, opts.Name == "athenz.hockey")
 	assert.Equal(t, opts.Profile, "test-profile")
 
-	assert.Equal(t, opts.Threshold, DefaultThreshold)
-	assert.Equal(t, opts.SshThreshold, DefaultThreshold)
+	assert.Equal(t, opts.Threshold, sc.DefaultThreshold)
+	assert.Equal(t, opts.SshThreshold, sc.DefaultThreshold)
 }
 
 // TestOptionsWithProfileConfig test the scenario when profile config file is present
@@ -258,12 +258,12 @@ func TestOptionsWithProfileConfig(t *testing.T) {
 	assert.Equal(t, "athenz.api", opts.Name)
 
 	// Zeroth service should be the one from "service" key, the remaining are from "services" in no particular order
-	assert.True(t, assertService(opts.Services[0], Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: DefaultThreshold}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: DefaultThreshold}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: DefaultThreshold}))
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: sc.DefaultThreshold}))
 
-	assert.Equal(t, DefaultThreshold, opts.SshThreshold)
-	assert.Equal(t, DefaultThreshold, opts.Threshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.SshThreshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.Threshold)
 	assert.Equal(t, "host1.athenz.io,host2.athenz.io", opts.SshPrincipals)
 	assert.False(t, opts.OmitDomain)
 
@@ -292,12 +292,12 @@ func TestOptionsWithProfileConfigAndProfileRestrictTo(t *testing.T) {
 	assert.True(t, opts.Name == "athenz.api")
 
 	// Zeroth service should be the one from "service" key, the remaining are from "services" in no particular order
-	assert.True(t, assertService(opts.Services[0], Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: DefaultThreshold}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: DefaultThreshold}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: DefaultThreshold}))
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: sc.DefaultThreshold}))
 
-	assert.Equal(t, DefaultThreshold, opts.SshThreshold)
-	assert.Equal(t, DefaultThreshold, opts.Threshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.SshThreshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.Threshold)
 }
 
 func TestOptionsWithRoleThreshold(t *testing.T) {
@@ -320,12 +320,12 @@ func TestOptionsWithRoleThreshold(t *testing.T) {
 	assert.Equal(t, "athenz.api", opts.Name)
 
 	// Zeroth service should be the one from "service" key, the remaining are from "services" in no particular order
-	assert.True(t, assertService(opts.Services[0], Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: 20}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: 20}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: 20}))
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: 20}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: 20}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: 20}))
 	assert.Equal(t, float64(25), opts.Roles[0].Threshold)
 
-	assert.Equal(t, DefaultThreshold, opts.SshThreshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.SshThreshold)
 	assert.Equal(t, float64(20), opts.Threshold)
 }
 
@@ -347,9 +347,9 @@ func TestOptionsWithServiceAccountThreshold(t *testing.T) {
 	assert.True(t, opts.Name == "athenz.api")
 
 	// Zeroth service should be the one from "service" key, the remaining are from "services" in no particular order
-	assert.True(t, assertService(opts.Services[0], Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: 35}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: 35}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: 25}))
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: 35}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: 35}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: 25}))
 
 	assert.Equal(t, float64(60), opts.SshThreshold)
 	assert.Equal(t, float64(35), opts.Threshold)
@@ -380,12 +380,12 @@ func TestOptionsWithConfig(t *testing.T) {
 	assert.Equal(t, "athenz.api", opts.Name)
 
 	// Zeroth service should be the one from "service" key, the remaining are from "services" in no particular order
-	assert.True(t, assertService(opts.Services[0], Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: DefaultThreshold}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: DefaultThreshold}))
-	assert.True(t, assertInServices(opts.Services[1:], Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: DefaultThreshold}))
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "ui", User: "root", Uid: 0, Gid: 0, FileMode: 288, Threshold: sc.DefaultThreshold}))
+	assert.True(t, assertInServices(opts.Services[1:], sc.Service{Name: "yamas", User: "nobody", Uid: getUid("nobody"), Group: "sys", Gid: getGid(t, "sys"), Threshold: sc.DefaultThreshold}))
 
-	assert.Equal(t, DefaultThreshold, opts.Threshold)
-	assert.Equal(t, DefaultThreshold, opts.SshThreshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.Threshold)
+	assert.Equal(t, sc.DefaultThreshold, opts.SshThreshold)
 }
 
 // TestOptionsNoService test the scenario when /etc/sia/sia_config is present, but service is not repeated in services
@@ -411,7 +411,7 @@ func TestOptionsNoServices(t *testing.T) {
 	assert.Equal(t, 1, len(opts.Services))
 	assert.Equal(t, "athenz", opts.Domain)
 	assert.Equal(t, "athenz.api", opts.Name)
-	assert.True(t, assertService(opts.Services[0], Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: DefaultThreshold}))
+	assert.True(t, assertService(opts.Services[0], sc.Service{Name: "api", User: "nobody", Uid: getUid("nobody"), Gid: getUserGid("nobody"), FileMode: 288, Threshold: sc.DefaultThreshold}))
 	assert.True(t, opts.FileDirectUpdate)
 }
 
@@ -564,7 +564,7 @@ func TestGetRunsAsUidGid(t *testing.T) {
 	}
 }
 
-func toServiceNames(services []Service) []string {
+func toServiceNames(services []sc.Service) []string {
 	var serviceNames []string
 
 	for _, srv := range services {
@@ -595,7 +595,7 @@ func TestOptionsWithAccessToken(t *testing.T) {
 		Service:                  "api",
 		Domain:                   "athenz.demo",
 		Roles:                    []string{"reader"},
-		Expiry:                   DefaultTokenExpiry,
+		Expiry:                   sc.DefaultTokenExpiry,
 		User:                     "nobody",
 		ProxyPrincipalSpiffeUris: "",
 	}))
@@ -615,7 +615,7 @@ func TestOptionsWithAccessToken(t *testing.T) {
 		Service:                  "api",
 		Domain:                   "athenz.demo",
 		Roles:                    []string{"reader", "reader-admin"},
-		Expiry:                   DefaultTokenExpiry,
+		Expiry:                   sc.DefaultTokenExpiry,
 		User:                     "nobody",
 		ProxyPrincipalSpiffeUris: "",
 	}))
@@ -635,7 +635,7 @@ func TestOptionsWithAccessToken(t *testing.T) {
 		Service:                  "logger",
 		Domain:                   "athenz.demo",
 		Roles:                    []string{"splunk"},
-		Expiry:                   DefaultTokenExpiry,
+		Expiry:                   sc.DefaultTokenExpiry,
 		User:                     "nobody",
 		ProxyPrincipalSpiffeUris: "",
 	}))
@@ -645,7 +645,7 @@ func TestOptionsWithAccessToken(t *testing.T) {
 		Service:                  "api",
 		Domain:                   "athenz.demo",
 		Roles:                    []string{"*"},
-		Expiry:                   DefaultTokenExpiry,
+		Expiry:                   sc.DefaultTokenExpiry,
 		User:                     "nobody",
 		ProxyPrincipalSpiffeUris: "",
 	}))
@@ -680,7 +680,7 @@ func TestInvalidAccessTokenDomain(t *testing.T) {
 }
 
 func TestGetRoleServiceOwner(t *testing.T) {
-	services := []Service{
+	services := []sc.Service{
 		{
 			Name: "svc1",
 		},
@@ -897,6 +897,6 @@ func TestOptionsWithServiceOnlySetup(t *testing.T) {
 	assert.Equal(t, opts.Name, "athenz.hockey")
 	assert.Equal(t, opts.Profile, "test-profile")
 
-	assert.Equal(t, opts.Threshold, DefaultThreshold)
-	assert.Equal(t, opts.SshThreshold, DefaultThreshold)
+	assert.Equal(t, opts.Threshold, sc.DefaultThreshold)
+	assert.Equal(t, opts.SshThreshold, sc.DefaultThreshold)
 }
