--- libs/go/sia/aws/sds/data/athenz.api.cert.pem
@@ -1,1 +1,0 @@
-x509-certificate
