--- libs/go/sia/aws/sds/data/athenz.api.key.pem
@@ -1,1 +1,0 @@
-private-key
