--- libs/go/sia/aws/sds/data/ca.cert.pem
@@ -1,1 +1,0 @@
-ca-certs
