+++ provider/gcp/sia-gce/cmd/siad/main.go
@@ -19,15 +19,16 @@ package main
 import (
 	"flag"
 	"fmt"
-	"github.com/AthenZ/athenz/libs/go/sia/ssh/hostkey"
-	"github.com/AthenZ/athenz/libs/go/sia/util"
 	"log"
 	"os"
 	"strings"
 
 	"github.com/AthenZ/athenz/libs/go/sia/agent"
+	sc "github.com/AthenZ/athenz/libs/go/sia/config"
 	"github.com/AthenZ/athenz/libs/go/sia/gcp/meta"
 	"github.com/AthenZ/athenz/libs/go/sia/options"
+	"github.com/AthenZ/athenz/libs/go/sia/ssh/hostkey"
+	"github.com/AthenZ/athenz/libs/go/sia/util"
 	"github.com/AthenZ/athenz/provider/gcp/sia-gce"
 )
 
@@ -99,7 +100,7 @@ func main() {
 	}
 
 	// backward compatibility sake, keeping the ConfigAccount struct
-	configAccount := &options.ConfigAccount{
+	configAccount := &sc.ConfigAccount{
 		Name:         fmt.Sprintf("%s.%s", config.Domain, config.Service),
 		User:         config.User,
 		Group:        config.Group,
