+++ provider/gcp/sia-gke/cmd/siad/main.go
@@ -24,6 +24,7 @@ import (
 	"strings"
 
 	"github.com/AthenZ/athenz/libs/go/sia/agent"
+	sc "github.com/AthenZ/athenz/libs/go/sia/config"
 	"github.com/AthenZ/athenz/libs/go/sia/gcp/meta"
 	"github.com/AthenZ/athenz/libs/go/sia/host/utils"
 	"github.com/AthenZ/athenz/libs/go/sia/options"
@@ -85,7 +86,7 @@ func main() {
 	}
 
 	// backward compatibility sake, keeping the ConfigAccount struct
-	configAccount := &options.ConfigAccount{
+	configAccount := &sc.ConfigAccount{
 		Name:         fmt.Sprintf("%s.%s", config.Domain, config.Service),
 		User:         config.User,
 		Group:        config.Group,
